legal_literacy_bridge/requirements.txt
streamlit>=1.35.0
requests>=2.31.0
pydub>=0.25.1
python-dotenv>=1.0.0
